+++ src/BrefServiceProvider.php
@@ -67,7 +67,13 @@ class BrefServiceProvider extends ServiceProvider
      */
     public function boot(Dispatcher $dispatcher, LogManager $logManager, FailedJobProviderInterface $queueFailer)
     {
+        $this->app->useStoragePath(StorageDirectories::Path);
+
         if ($this->app->runningInConsole()) {
+            $this->publishes([
+                __DIR__ . '/../stubs/serverless.yml' => config_path('serverless.yml'),
+            ], 'serverless-config');
+
             $this->publishes([
                 __DIR__ . '/../config/bref.php' => config_path('bref.php'),
             ], 'bref-config');